+++ sharding-orchestration/sharding-orchestration-reg/src/main/java/io/shardingsphere/orchestration/reg/newzk/client/zookeeper/strategy/SyncRetryStrategy.java
@@ -15,12 +15,13 @@
  * </p>
  */
 
-package io.shardingsphere.jdbc.orchestration.reg.newzk.client.zookeeper.strategy;
+package io.shardingsphere.orchestration.reg.newzk.client.zookeeper.strategy;
 
-import io.shardingsphere.jdbc.orchestration.reg.newzk.client.action.IProvider;
-import io.shardingsphere.jdbc.orchestration.reg.newzk.client.retry.DelayRetryPolicy;
-import io.shardingsphere.jdbc.orchestration.reg.newzk.client.retry.RetryCallable;
-import io.shardingsphere.jdbc.orchestration.reg.newzk.client.retry.RetryResultCallable;
+
+import io.shardingsphere.orchestration.reg.newzk.client.action.IProvider;
+import io.shardingsphere.orchestration.reg.newzk.client.retry.DelayRetryPolicy;
+import io.shardingsphere.orchestration.reg.newzk.client.retry.RetryCallable;
+import io.shardingsphere.orchestration.reg.newzk.client.retry.RetryResultCallable;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
